+++ js/validation.js
@@ -1,0 +1,5 @@
+//Функция проверки максимальной длины строки
+
+const checkLength = (string, maxLength) => string.length <= maxLength;
+
+checkLength ('Пример строки', 15);
